+++ server.js
@@ -7,8 +7,7 @@ require('dotenv').config();
 require('./config/database');
 
 const indexRouter = require('./routes/index');
-const cropRouter = require('./routes/crops');
-// const detailsRouter= require('./routes/details');
+const itemRouter = require('./routes/items');
 
 var app = express();
 
@@ -23,8 +22,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/crops', cropRouter);
-// app.use('/', detailsRouter);
+app.use('/items', itemRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
