+++ views/index.ejs
@@ -8,10 +8,10 @@
     <title>Welcome!</title>
 </head>
   <body>
-    <a href="/crops">
+    <a href="/items">
         Shop!
         </a>
-    <a href="/crops/new">
+    <a href="/items/new">
     Sell!</a>
 </body>
 </html>
