+++ views/items/new.ejs
@@ -8,7 +8,7 @@
   </head>
   <body>
 <h2>Enter a new item listing</h2>
-<form action="/crops" method="POST">
+<form action="/items" method="POST">
     <label>
          Full Name: <input placeholder="Example: John Doe" type="text" class="name" autocomplete="name">
     </label><br>
@@ -19,13 +19,13 @@
      Primary Phone: <input placeholder="1234567891" type="number" class="phone" autocomplete="tel-national">
     </label> <br>
   <label>Type of Item:
-    <select name="cropType">
-              <option name="Fruits" value="Fruits">Fruits</option>
-              <option name="Vegetables" value="Vegetables">Vegetables</option>
+    <select name="itemType">
+              <option value="Fruits">Fruits</option>
+              <option value="Vegetables">Vegetables</option>
     </select>
   </label>
 	<label>Item Type:
-	  <input type="text" class="custom-fruit" name="name" placeholder="Enter an Item">
+	  <input type="text" class="custom-item" name="name" placeholder="Enter an Item">
 	</label><br>
 	<label>Quantity:
 	   <select name="quantity">
