+++ views/items/show.ejs
@@ -9,13 +9,13 @@
 <body>
      <section id="show-page">
         <div>Fruits: </div>
-        <div><%= cropType.fruits %></div>
+        <div><%= item.fruits %></div>
         <div>Vegetables: </div>
-        <div><%= cropType.vegetables%></div>
+        <div><%= item.vegetables%></div>
         <div>Quantity: </div>
-        <div><%= crop.quantity %></div>
+        <div><%= item.quantity %></div>
         <div>Price: </div>
-        <div><%= crop.price%></div>
+        <div><%= item.price%></div>
     </section>
 </body>
 </html>
